requirements.txt
Django>=5.1
Flask>=3.0
SQLAlchemy>=2.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-CORS>=4.0
Werkzeug>=3.0
numpy>=2.1
pandas>=2.2
scikit-learn>=1.5
transformers>=4.44
torch>=2.4
torchvision>=0.19
opencv-python>=4.10
Pillow>=10.3
matplotlib>=3.9
seaborn>=0.13
requests>=2.32
python-dotenv>=1.0